pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "mcp>=1.26.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
